test: Restructure tests and add a case for global serializers
