+++ tests/test_elements.py
@@ -1863,11 +1863,7 @@ class TestSerializers:
     def test_add_attributes(self, model: capellambse.MelodyModel):
         converters = {
             "add_attributes": [
-                {
-                    "capella_attr": "layer",
-                    "polarion_id": "layer",
-                    "enum_id": "layer",
-                },
+                {"capella_attr": "layer", "polarion_id": "layer"},
                 {"capella_attr": "nature", "polarion_id": "nature"},
                 {"capella_attr": "kind", "polarion_id": "kind"},
             ]
@@ -2145,41 +2141,13 @@ class TestSerializers:
         assert work_item == data_model.CapellaWorkItem(**expected)
 
     @staticmethod
-    def test_read_config_context_diagram_with_params():
-        expected_filter = (
-            "capellambse_context_diagrams-show.exchanges.or.exchange.items."
-            "filter"
-        )
-        config = converter_config.ConverterConfig()
-        with open(TEST_MODEL_ELEMENTS_CONFIG, "r", encoding="utf8") as f:
-            config.read_config_file(f)
-
-        type_config = config.get_type_config("sa", "SystemFunction")
-
-        assert type_config is not None
-        assert isinstance(type_config.converters, dict)
-        assert "add_context_diagram" in type_config.converters
-        assert type_config.converters["add_context_diagram"]["filters"] == [
-            expected_filter
-        ]
-
-    @staticmethod
-    def test_read_config_tree_view_with_params(
-        model: capellambse.MelodyModel,
-    ):
+    def test_tree_view_with_params(model: capellambse.MelodyModel):
         cap = model.by_uuid("c710f1c2-ede6-444e-9e2b-0ff30d7fd040")
         config = converter_config.ConverterConfig()
         with open(TEST_MODEL_ELEMENTS_CONFIG, "r", encoding="utf8") as f:
             config.read_config_file(f)
 
         type_config = config.get_type_config("la", "Class")
-        assert type_config is not None
-        assert isinstance(type_config.converters, dict)
-        assert "add_tree_diagram" in type_config.converters
-        assert type_config.converters["add_tree_diagram"]["render_params"] == {
-            "depth": 1
-        }
-
         serializer = element_converter.CapellaWorkItemSerializer(
             model,
             polarion_repo.PolarionDataRepository(),
@@ -2199,47 +2167,6 @@ class TestSerializers:
 
         assert wrapped_render.call_count == 1
         assert wrapped_render.call_args_list[0][1] == {"depth": 1}
-
-    @staticmethod
-    def test_read_config_links(caplog: pytest.LogCaptureFixture):
-        caplog.set_level("DEBUG")
-        config = converter_config.ConverterConfig()
-        expected = (
-            "capella2polarion.converters.converter_config",
-            20,
-            "Global link parent is not available on Capella type diagram",
-            "capella2polarion.converters.converter_config",
-            40,
-            "Link exchanged_items is not available on Capella type "
-            "FunctionalExchange",
-        )
-        with open(TEST_MODEL_ELEMENTS_CONFIG, "r", encoding="utf8") as f:
-            config.read_config_file(f)
-
-        assert config.diagram_config
-        assert not any(
-            link
-            for link in config.diagram_config.links
-            if link.capella_attr == "parent"
-        )
-        assert caplog.record_tuples[0] + caplog.record_tuples[1] == expected
-        assert (
-            system_fnc_config := config.get_type_config("sa", "SystemFunction")
-        )
-        assert system_fnc_config.links[0] == converter_config.LinkConfig(
-            capella_attr="inputs.exchanges",
-            polarion_role="input_exchanges",
-            include={"Exchange Items": "exchange_items"},
-            link_field="inputExchanges",
-            reverse_field="inputExchangesReverse",
-        )
-        assert system_fnc_config.links[1] == converter_config.LinkConfig(
-            capella_attr="outputs.exchanges",
-            polarion_role="output_exchanges",
-            include={"Exchange Items": "exchange_items"},
-            link_field="output_exchanges",
-            reverse_field="output_exchanges_reverse",
-        )
 
     @staticmethod
     def test_add_context_diagram_with_caption(model: capellambse.MelodyModel):
